requirement.txt
dnspython
bs4
requests
PyPDF2